Refactor imports and API paths
